RDKB-5981 : GPLv2 match found in hotspot
Reason for change: snoop_udpChecksum is not required as UDP checksum is not added in any of the DHCP packets
Test Procedure: Enable Xfinitywifi and connect some clients to xfinitywifi 2.4GHZ and 5GHZ radios.
Risks: Xfinitywifi 2.4GHZ and 5GHZ clients should be able to connect, get the ip and browse.
Signed-off-by: Raghavendra Kammara <raghavendra_kammara@cable.comcast.com>

Change-Id: I157042c27ddbb8017b13aa494d32a87bd07fdb83