RDKCOM-2179: RDKBDEV-419: fix misuse of "\0"
Reason for change: "\0" is a pointer to an empty string (ie effectively the same as ""),
and is NOT the same as a nul charactor (ie a char with value 0).
Usage of "\0" is almost always a bug. If a pointer to an empty string
is needed then use "", if a nul character is needed then use 0 or
'\0'.
Test Procedure: Build and Sanity
Risks: None.
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

(cherry picked from commit b810501087ec4bfe373a8f352660aa31071b300c)
(cherry picked from commit 5e4d9e9df9707893f308ef78bf5255d78ba7fedb)
Change-Id: I38c9b6946716133e3d1a0443a185bcb62b004934