RDKBDEV-419: fix misuse of "\0"
Reason for change: "\0" is a pointer to an empty string (ie effectively the same as ""),
and is NOT the same as a nul charactor (ie a char with value 0).

Usage of "\0" is almost always a bug. If a pointer to an empty string
is needed then use "", if a nul character is needed then use 0 or
'\0'.

Test Procedure: Build and Sanity
Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: Ic1eec4339d5540164cbe8f929552f25f8f69cd86