fix gVlanSyncData queue_num values for generic / non-Comcast case ( OFW-5928 )
Fix CcspHotspot_process() crash for generic / non-Comcast builds.

[BUG DESCRIPTION]
During ccspHotspot process initialization, it will create queue handle for processing packets using the function nfq_create_queue
But after rdk rebase the queue number is not passed properly, because of this handle is not created and this will exit the process.

[FIX DESCRIPTION]
Done code changes for passing the queue number number properly

[Testing]
1.Enable community wifi from boot file.
2.Using cdrouter setup confirmed SSID is broadcasting properly and connectivity works fine.
3.Also checked /proc/net/netfilter/nfnetlink_queue and confirmed that handle is created with ccsphotspot process

Signed-off-by: Delvin C Varghese <dvarghese.contractor@libertyglobal.com>